[#49] refactor: 로그인&검색 컴포넌트 분리 및 클라이언트 컴포넌트 최소화
